backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses==0.6
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-mptt==0.16.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djongo==1.2.31
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.70.0
pillow==11.1.0
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
